Got manual control working

+++ src/main/java/frc/robot/StateMachine/MechStates/MechState.java
@@ -1,5 +1,5 @@
 package frc.robot.StateMachine.MechStates;
 
 public interface MechState {
-    int getPosition();
+    double getPosition();
 }
